requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
sentence-transformers>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
psutil>=5.9.0
httpx>=0.24.0
prometheus-fastapi-instrumentator>=6.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
